remove test request file
